apps/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy[asyncio]
alembic
asyncpg
redis
rq
python-jose
passlib[argon2]
httpx
meilisearch
weasyprint
orjson
structlog
pillow
pytest
pytest-asyncio
httpx
